+++ .pre-commit-config.yaml
@@ -1,0 +1,65 @@
+repos:
+  - repo: https://github.com/pre-commit/pre-commit-hooks
+    rev: v4.6.0
+    hooks:
+      - id: check-merge-conflict
+        stages: [ pre-commit, pre-push ]
+      - id: check-ast
+        stages: [ pre-commit ]
+      - id: trailing-whitespace
+        stages: [ pre-commit ]
+      - id: end-of-file-fixer
+        stages: [ pre-commit ]
+
+  - repo: local
+    hooks:
+      - id: uv
+        name: uv-lock
+        description: "Run uv lock"
+        entry: uv lock --locked
+        language: system
+        pass_filenames: false
+        types: [ python ]
+        stages: [ pre-commit ]
+
+      - id: pyright
+        name: pyright
+        description: "Run 'pyright' for Python type checking"
+        entry: uv run pyright fntypes --level=ERROR
+        pass_filenames: false
+        language: system
+        types: [ python ]
+        stages: [ pre-commit ]
+
+      - id: pytest
+        name: pytest
+        description: "Run tests for testing code"
+        entry: uv run pytest tests
+        language: system
+        pass_filenames: false
+        types: [ python ]
+        stages: [ pre-commit ]
+
+      - id: ruff
+        name: ruff-isort
+        description: "Run 'ruff --select I --select F401 --fix' for extremely fast Python sorting imports"
+        entry: uv run ruff check --select I --select F401 --fix
+        language: system
+        types: [ python ]
+        stages: [ pre-commit ]
+
+      - id: ruff
+        name: ruff-format
+        description: "Run 'ruff format' for extremely fast Python formatting"
+        entry: uv run ruff format
+        language: system
+        types: [ python ]
+        stages: [ pre-commit ]
+
+      - id: ruff
+        name: ruff-check
+        description: "Run 'ruff-format' for extremely fast Python linting"
+        entry: uv run ruff check --fix
+        language: system
+        types: [ python ]
+        stages: [ pre-commit ]
