+++ fntypes/library/__init__.py
@@ -1,12 +1,11 @@
-from fntypes.library.caching import cache, acache
+from fntypes.library.caching import acache, cache
 from fntypes.library.error import UnwrapError
 from fntypes.library.functor import F
-from fntypes.library.misc import this, either, is_ok, is_err, is_some, is_nothing, from_optional
-from fntypes.library.variative import Variative
-from fntypes.library.monad import Result, Ok, Error, Option, Some, Nothing
-from fntypes.library.unwrapping import unwrapping
 from fntypes.library.lazy import Lazy, LazyCoro, LazyCoroResult
-
+from fntypes.library.misc import either, from_optional, is_err, is_nothing, is_ok, is_some, this
+from fntypes.library.monad import Error, Nothing, Ok, Option, Result, Some
+from fntypes.library.unwrapping import unwrapping
+from fntypes.library.variative import Variative
 
 __all__ = (
     "cache",
