+++ fntypes/library/lazy/__init__.py
@@ -2,7 +2,6 @@ from fntypes.library.lazy.lazy import Lazy
 from fntypes.library.lazy.lazy_coro import LazyCoro
 from fntypes.library.lazy.lazy_coro_result import LazyCoroResult
 
-
 __all__ = (
     "Lazy",
     "LazyCoro",
