+++ fntypes/library/lazy/lazy_coro.py
@@ -2,8 +2,8 @@ from __future__ import annotations
 
 import typing
 
-from fntypes.library.lazy.lazy import Lazy
 from fntypes.library.caching import acache
+from fntypes.library.lazy.lazy import Lazy
 
 
 class LazyCoro[Value]:
