+++ fntypes/library/monad/__init__.py
@@ -1,5 +1,5 @@
-from fntypes.library.monad.result import Result, Ok, Error
-from fntypes.library.monad.option import Option, Some, Nothing
+from fntypes.library.monad.option import Nothing, Option, Some
+from fntypes.library.monad.result import Error, Ok, Result
 
 __all__ = (
     "Result",
