+++ fntypes/utilities/misc.py
@@ -1,9 +1,24 @@
 from __future__ import annotations
+
+import inspect
 import types
 import typing
 
-Instance = typing.TypeVar("Instance")
-Attribute = typing.TypeVar("Attribute")
+type Caster[T, R] = typing.Callable[[T], R] | type[type[T]]
+
+
+def is_dunder(attr: str) -> bool:
+    return attr.startswith("__") and attr.endswith("__")
+
+
+def get_frame(depth: int = 0) -> types.FrameType | None:
+    frame = inspect.currentframe()
+
+    for _ in range(depth):
+        if frame is not None:
+            frame = frame.f_back
+
+    return frame
 
 
 class _GetArgsDescriptor[Instance, Attribute]:
@@ -19,10 +34,3 @@ class BindStaticMeta(type):
             namespace["get_args"] = _GetArgsDescriptor()
 
         return super().__new__(mcls, name, bases, namespace)
-
-
-def is_dunder(attr: str) -> bool:
-    return attr.startswith("__") and attr.endswith("__")
-
-
-type Caster[T, R] = typing.Callable[[T], R] | type[type[T]]
