+++ pyproject.toml
@@ -13,6 +13,7 @@ dev = [
     "pytest-asyncio>=1.0.0,<2",
     "ruff>=0.11.12,<0.12",
     "pyright>=1.1.401,<2",
+    "pre-commit>=4.2.0,<5.0.0"
 ]
 
 [build-system]
